Qemu 2: apply ML patchset, port archive-source.sh
Later Qemu versions have a convenience script that
collects the required files to build in an isolated environment,
this was easy to backport, letting docker build scripts
use it for both versions.

+++ vl.c
@@ -4102,7 +4102,7 @@ int main(int argc, char **argv, char **envp)
     }
 
     if (log_mask) {
-        int mask;
+        uint64_t mask;
         mask = qemu_str_to_log_mask(log_mask);
         if (!mask) {
             qemu_print_log_usage(stdout);
